Change class name
